% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonStandardEval.R
\name{bquote_remask}
\alias{bquote_remask}
\title{Automatically re-evaluates a bquote call in a local (or specified) environment}
\usage{
bquote_remask(expr, env = parent.frame())
}
\arguments{
\item{expr}{Either a bquote call, or an argument to be passed to bquote.}

\item{env}{an environment}
}
\description{
Also checks if the argument is already a call  to bquote. If not, then wraps the argument in bquote.
}
