% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{col_flatten_alpha}
\alias{col_flatten_alpha}
\title{Flatten alpha transparency onto a background color}
\usage{
col_flatten_alpha(color, alpha = 1, bg = "white")
}
\arguments{
\item{color}{Color value. String or Hex.}

\item{alpha}{Alpha. Value from 0 to 1.}

\item{bg}{Background color to flatten transparency onto. Default = "white".}
}
\description{
Flatten alpha transparency onto a background color
}
