% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{col_ramp_brewer}
\alias{col_ramp_brewer}
\alias{col_rmp_red}
\alias{col_rmp_blue}
\title{Color ramp shortcuts}
\usage{
col_ramp_brewer(palette = rownames(RColorBrewer::brewer.pal.info), alpha = 0.4)

col_rmp_red(x)

col_rmp_blue(x)
}
\arguments{
\item{palette}{An RColorBrewer palette.}

\item{alpha}{Transparency. Default 0.4.}

\item{x}{Argument to the returned object, which is a function that maps (0,1) -> hex color code.}
}
\description{
Quick shortcuts for RColorBrewer color ramp palettes that map \verb{[0,1]} to colors
}
