% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrixOperations.R
\name{combine_block_diagonal}
\alias{combine_block_diagonal}
\title{Combine two matrices along block diagonals}
\usage{
combine_block_diagonal(A, B, fill = NA)
}
\arguments{
\item{A}{Matrix x}

\item{B}{Matrix y}

\item{fill}{Fill value}
}
\description{
\verb{(A, B) -> [A 0; 0 B]}
}
