% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filesave.R
\name{file_datestamp}
\alias{file_datestamp}
\title{Append a datestamp to a filename}
\usage{
file_datestamp(
  input_filename = stop("Input filename(s) required"),
  path = getwd()
)
}
\arguments{
\item{input_filename}{Input filename(s) to check, as characters}

\item{path}{Optional path to prepend to filename. Defaults to current working directory. If a relative path is provided, path is assumed relative to the current working directory.}
}
\value{
A modified filepath
}
\description{
Will also append a counter to prevent overwriting existing files
}
