% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_save.R
\name{flextable_save}
\alias{flextable_save}
\title{Save a flextable to disk in several formats}
\usage{
flextable_save(data, name = waiver(), path = waiver(), overwrite = FALSE)
}
\arguments{
\item{data}{Input flextable}

\item{name}{Output file name}

\item{path}{Output file path}

\item{overwrite}{Whether to overwrite an existing file. Default = FALSE, which appends a numeric index to the filename.}
}
\value{
The original flextable, unmodified
}
\description{
Save a flextable to disk as an html, rds (compressed), and tsv.gz (containing the dataframe)
}
