% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_scibeautify.R
\name{flextable_scibeautify}
\alias{flextable_scibeautify}
\title{Apply scibeautify to a flextable}
\usage{
flextable_scibeautify(data, .across = c(), ...)
}
\arguments{
\item{data}{Input data frame}

\item{.across}{Tidy-eval selector to apply scibeautify to}

\item{...}{
  Arguments passed on to \code{\link[=scibeautify]{scibeautify}}
  \describe{
    \item{\code{sci_mode}}{(Optional) Can be one of "auto" (automatically choose whether scientific notation is used based on string length optimization), "on", or "off" (scientific notation always on or off).}
    \item{\code{justify_mode}}{Mode to use for justifying a numbers. Can be "l" (left), "c" (center), "r" (right), or "d" (decimal). An "x" flag can be appended which will independently align the exponent portion of numbers in scientific notation, in addition to aligning the coefficient. Allowed values = \code{c("l", "c", "r", "d", "lx", "cx", "rx", "dx")}.}
    \item{\code{auto_ignore_int}}{(Optional) If the data (or column) is all integers, will not truncate non-significant digits and will not use scientific notation}
    \item{\code{.cols}}{Tidy selector for columns when \code{.data} is a dataframe. Default = \code{tidyr::everything()}.}
    \item{\code{.rows}}{Vector of row numbers indicating where scibeautify should be applied.}
    \item{\code{sig_digits}}{Number of significant digits. If NA or NULL, the input number is not changed.}
    \item{\code{pad_zeros}}{Whether to add extra trailing zeros and/or a decimal to force number of significant digits (even if the original number did not have that many digits). Default = FALSE.}
    \item{\code{decimal_mark}}{Character to use as the decimal marker.}
    \item{\code{big_mark}}{If not blank, character to use in grouping every \code{big_interval} digits before the decimal marker.}
    \item{\code{big_interval}}{Number of digits between each \code{big_mark}, if used.}
    \item{\code{nsmall}}{Forces output to have nsmall digits after the decimal. If there are fewer significant digits than total digits after including nsmall, the number of significant digits is increased.}
  }}
}
\value{
A flextable
}
\description{
This is a wrapper around both \code{\link[flextable:flextable]{flextable::flextable()}} and \code{\link[=scibeautify]{scibeautify()}} that applies \code{\link[flextable:as_sup]{flextable::as_sup()}} where needed to scientific notation superscripts.
}
