% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_beautify.R
\name{format_signif}
\alias{format_signif}
\title{Convert a number to a string in significant-figure form}
\usage{
format_signif(n, sig_digits = 3, padzeros = T, justify = F)
}
\arguments{
\item{n}{Number or numeric vector to convert}

\item{sig_digits}{(Optional) Number of significant digits (default = 3)}

\item{padzeros}{(Optional) Whether to add trailing zeros (default = TRUE)}

\item{justify}{(Optional) Whether to left-justify string output across a character vector (default = FALSE)}
}
\description{
Convert a number to a string in significant-figure form
}
