% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{Half-violin plot
From https://www.data-to-viz.com/caveat/boxplot.html}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
Half-violin plot
From https://www.data-to-viz.com/caveat/boxplot.html
}
