% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_halfViolin.R
\name{geom_half_violin}
\alias{geom_half_violin}
\title{Half-violin plot
From https://www.data-to-viz.com/caveat/boxplot.html}
\usage{
geom_half_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat = "ydensity"}

\item{position}{position = "dodge"}

\item{trim}{trim = TRUE}

\item{scale}{scale = "area"}

\item{show.legend}{show.legend = NA}

\item{inherit.aes}{inherit.aes = TRUE}

\item{...}{other ggplot layer params}
}
\description{
Half-violin plot
From https://www.data-to-viz.com/caveat/boxplot.html
}
