% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_concat.R
\name{gg_concat}
\alias{gg_concat}
\title{Concatenate any ggplot components}
\usage{
gg_concat(e1, e2)
}
\arguments{
\item{e1}{Left-hand side of addition operation}

\item{e2}{Right-hand side of addition operation}
}
\value{
A list of gg objects
}
\description{
Typically, addition in \code{ggplot2} with the \code{+} operator is only permitted when the left-hand side is a \code{ggplot} object or a \code{theme} object. This function extends the ggplot2 functionality and allows the user to concatenate geom, theme, stats, labels, aes, Coord, and Scale objects into bundles using the \code{+} operator. These compositions can then be stored and used later in the construction of ggplots.
}
\details{
A common use-case for this is to construct custom plot templates by concatenating base geoms and settings.

The addition function is designed to capture all gg objects that are not processed by the original +.gg function. See \href{https://github.com/tidyverse/ggplot2/blob/13c07309e5d6798277d5d0c6f76ead88b6112e40/R/plot-construction.r#L42}{original source}
}
\section{Usage}{

custom_geom <- function(fill = "red") {
aes(fill = !!fill) +
geom_boxplot() +
geom_jitter()
}

ggplot() + custom_geom()
}

