% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_glmforest.R
\name{gg_glmforest}
\alias{gg_glmforest}
\title{Forest plot for generalized linear models}
\usage{
gg_glmforest(
  glm_list,
  stats.predictor_var,
  stats.conf_interval = 0.95,
  stats.effect_sig_digits = 3,
  stats.sig_pvalue = waiver(),
  label.effect_size = paste("OR per unit", stats.predictor_var),
  label.dependent_var = "Dependent Var.",
  plot.rel_width = 0.33,
  plot.point.size = 1.8,
  plot.point.color = "dodgerblue4",
  plot.row.color = col_flatten_alpha(plot.point.color, 0.06),
  plot.breaks = waiver(),
  plot.vertical.pos = waiver(),
  plot.vertical.label = waiver()
)
}
\arguments{
\item{glm_list}{A named list of the results of calls to \code{lm()} or \code{glm()}. Names will be used as labels in the table.}

\item{stats.predictor_var}{The name of the variable in the linear model that is being regressed on (i.e., the predictor variable).}

\item{stats.conf_interval}{Confidence interval. Default = 95\%.}

\item{stats.effect_sig_digits}{Number of significant figures to for effect size and confidence interval}

\item{stats.sig_pvalue}{(Not used) significant p-value after correction for multiple comparisons.}

\item{label.effect_size}{Label/column header describing the effect size.}

\item{label.dependent_var}{Label/column header for the dependent variables (i.e., predicted variables).}

\item{plot.rel_width}{Width of the forest plot relative to the width of the table. Default = 0.33.}

\item{plot.point.size}{Size of the boxes in the forest plot. Default = 1.8.}

\item{plot.point.color}{Color of the boxes in the forest plot. Default = "dodgerblue4".}

\item{plot.row.color}{Background color of alternating rows of the table. Default = "dodgerblue4" with 4\% transparency. Use NA to omit background colors.}

\item{plot.breaks}{Breaks to use on the horizontal axis of the forest plot.}

\item{plot.vertical.pos}{Vertical position of the indicator line on the forest plot. Default = 1 for logistic regression, 0 otherwise. Use NA to omit.}

\item{plot.vertical.label}{Boolean, whether to add a label on the horizontal axis corresponding to the vertical indicator, if not already included in the breaks.}
}
\description{
Generate a forest plot with a table, optimized for logistic regression.
}
