% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_manhattan.R
\name{gg_manhattan}
\alias{gg_manhattan}
\title{Manhattan plot with ggplot2}
\usage{
gg_manhattan(
  data,
  chr,
  pos,
  pval,
  label,
  title = NULL,
  label.mode = c("none", "fdr05", "signif", "suggest"),
  label.size = 0.7,
  point.size = 0.8,
  hline.signif = NA,
  hline.suggest = NA,
  highlight_loci.thresh = hline.signif,
  highlight_loci.width = 250,
  highlight_loci.color = "red",
  overplot.limit = 0.05,
  overplot.frac = 0.05,
  chr.colors = manhattan_colors_GnBu(),
  chr.labels = waiver(),
  limits_y = waiver(),
  breaks_y = waiver(),
  progress = TRUE
)
}
\arguments{
\item{data}{Input data frame}

\item{chr}{Name of column with chromosome number.}

\item{pos}{Name of column with chromosome position (typically in base pair)}

\item{pval}{Name of column with p-value}

\item{label}{Optional, name of column with rsid symbols, gene symbols, or other labels}

\item{title}{Plot title}

\item{label.mode}{How to apply labels to selected points. One of \code{c("none", "fdr_0.05", "hline.signif", "hline.suggest")}. \code{"fdr_0.05"} identifies points that have FDR < 0.05. \code{"hline.signif"} identifies points above the significance line.}

\item{label.size}{Font size of labels relative to the ggplot theme text element}

\item{point.size}{Size of point. Default 0.8.}

\item{hline.signif}{Horizontal significance line. Can be a number, or "bonferroni" to automatically calculate Bonferroni correction. Use NA to omit.}

\item{hline.suggest}{Horizontal suggestive line. Can be a number, or "bonferroni" to automatically calculate Bonferroni correction. Use NA to omit.}

\item{highlight_loci.thresh}{Threshold p-value to use to identify lead snps in loci that will be highlighted with highlight_loci.color. Use NA to omit.}

\item{highlight_loci.width}{Width, in kilobases, of loci identified by highlight_loci.thresh to include in the highlight regions}

\item{highlight_loci.color}{Highlight color}

\item{overplot.limit}{If there are greater than 1 million points, only a fraction (specified by \code{overplot.frac}) of the points with p-values above this limit will be randomly removed.}

\item{overplot.frac}{Fraction of points above the overplot.limit to plot. Other points (randomly sampled) will be discarded.}

\item{chr.colors}{Vector of color names, applied to each chromosome.}

\item{chr.labels}{Optional. Chromosome labels.}

\item{limits_y}{Optional. y-axis limits, in -log10 units.}

\item{breaks_y}{Optional. y-axis breaks, in -log10 units.}

\item{progress}{Whether to show a progress bar during ggplot construction}
}
\description{
...
}
