% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_figsave.R
\name{ggfigsave}
\alias{ggfigsave}
\title{Output and save a ggplot object}
\usage{
ggfigsave(
  ggdata,
  name = waiver(),
  path = waiver(),
  dpi = 600,
  width = 6.5,
  height = waiver(),
  aspect_ratio = 1,
  units = c("in", "cm", "mm", "px"),
  format = c("pdf", "png", "rds"),
  open = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{ggdata}{ggplot2 object}

\item{name}{Filename (optional, defaults to name of the input ggplot object)}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Only applies when converting pixel units,
as is typical for raster output types.}

\item{width, height}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}

\item{aspect_ratio}{Ratio of width to height. Default = 1.}

\item{units}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}

\item{format}{Output file format, any of "rds", "pdf", "eps", "ps", "tex", "jpeg", "tiff", "png", "bmp", "svg".}

\item{open}{Open the image file after creation. Default = TRUE.}

\item{overwrite}{Whether to overwrite an existing file. Default = FALSE, which appends a numeric index to the filename.}
}
\value{
Original ggplot object, invisibly.
}
\description{
Outputs a ggplot object as a pdf with standard scientific journal settings, and saves and rds object
}
