% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_bland.altman.R
\name{ggplot_bland.altman}
\alias{ggplot_bland.altman}
\title{Generate a Bland-Altman plot}
\usage{
ggplot_bland.altman(
  method1,
  method2,
  fill.colors = c("steelblue3", "gray85"),
  ...
)
}
\description{
Generate a Bland-Altman plot
}
