% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_qqplot_aj.R
\name{ggplot_qqplot_aj}
\alias{ggplot_qqplot_aj}
\title{Generate a QQ-plot with confidence bands}
\usage{
ggplot_qqplot_aj(
  data,
  rescale.expected = FALSE,
  stat = stats::qnorm,
  ci = 0.95,
  label_x = "Expected",
  label_y = "Observed"
)
}
\arguments{
\item{data}{...}

\item{rescale.expected}{...}

\item{stat}{...}

\item{ci}{...}

\item{label_x}{...}

\item{label_y}{...}
}
\description{
UNDER DEVELOPMENT
}
