% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{invertHexColor}
\alias{invertHexColor}
\title{Invert a hex color}
\usage{
invertHexColor(hexColor)
}
\arguments{
\item{hexColor}{Hex color}
}
\value{
Hex color string, suitable for use in R graphics
}
\description{
Invert a hex color
}
