% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{is_integer_num}
\alias{is_integer_num}
\title{Is the number an integer?}
\usage{
is_integer_num(n)
}
\arguments{
\item{n}{Numeric to test}
}
\description{
Different from \code{\link[=is.integer]{is.integer()}} and \code{\link[rlang:type-predicates]{rlang::is_integer()}}, which only check if the class of the object is \code{"integer"}.
}
