% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_justify.R
\name{justify}
\alias{justify}
\title{Justify a character vector}
\usage{
justify(
  input_vec,
  justify_mode = c("decimal", "left", "center", "right"),
  decimal = ".",
  justify_char = " "
)
}
\arguments{
\item{input_vec}{Input vector. Coerced to character.}

\item{justify_mode}{Mode to use for justifying a vector of values. Can be "l" (left), "c" (center), "r" (right), or "d" (decimal).}

\item{decimal}{Character to justify around when justify_mode = "decimal"}

\item{justify_char}{Character (or string) used to pad strings.}
}
\description{
Justify a character vector of numbers using a character to pad strings on the left and right sides
}
