% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_LD_matrix.R
\name{ld_list_to_matrix}
\alias{ld_list_to_matrix}
\title{Convert a paired list of LD values to a symmetric matrix}
\usage{
ld_list_to_matrix(
  data,
  SNP_A = "SNP_A",
  SNP_B = "SNP_B",
  R = "R",
  missing_value = NA
)
}
\arguments{
\item{data}{Input dataframe}

\item{SNP_A}{Column name containing names of SNP_A}

\item{SNP_B}{Column name containing names of SNP_B}

\item{R}{Column name containing R (or R2) values}

\item{missing_value}{Value to use to fill missing entries}
}
\value{
Symmetric matrix
}
\description{
Convert a dataframe that lists pairs of snps with their LD values into a symmetric LD matrix
}
