% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_LD_matrix.R
\name{ld_matrix_LDLink}
\alias{ld_matrix_LDLink}
\title{Retrieve LD matrix with LDLink}
\usage{
ld_matrix_LDLink(
  snps,
  pop = "EUR",
  r2d = "r2",
  tokens = NULL,
  cache = NULL,
  cache_tokens = TRUE,
  max_query_size = 1000,
  progress = TRUE,
  compress = FALSE
)
}
\arguments{
\item{snps}{List of variants, using an rsID or chromosome coordinate (e.g. "chr7:24966446")}

\item{pop}{A 1000 Genomes Project population, default = "EUR"}

\item{r2d}{Either "r2" for LD R2 or "d" for LD D', default = "r2"}

\item{tokens}{List of LDLink tokens. If multiple tokens are provided, look-ups will be performed in parallel}

\item{cache}{Name of an optional cache file to read/store results from/to}

\item{cache_tokens}{Whether to save tokens in the exported LD cache file.}

\item{max_query_size}{Max number of SNPs to pass to LDlinkR at a time}

\item{progress}{Display a progress indicator}

\item{compress}{Whether output rds should be compressed}
}
\description{
This function is a wrapper around \code{\link[LDlinkR:LDmatrix]{LDlinkR::LDmatrix()}} that allows queries for greater than 1000 SNPs, and using an optional local cache file for faster look-ups. The LDlink API only allows queries of up to 1000s SNPs at a time. This function makes repeated calls to the API to build the appropriate LD matrix.
}
