% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_manhattan.R
\name{manhattan_colors_GnBu}
\alias{manhattan_colors_GnBu}
\alias{manhattan_colors_Crimson}
\title{Color ramp for Manhattan plots}
\usage{
manhattan_colors_GnBu()

manhattan_colors_Crimson()
}
\description{
Color ramp for Manhattan plots
}
