% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrixOperations.R
\name{mutate_rows}
\alias{mutate_rows}
\title{Apply dplyr's mutate to a subset of rows}
\usage{
mutate_rows(.data, row_selector, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{row_selector}{vector of row indices}

\item{...}{
  Arguments passed on to \code{\link[dplyr:mutate]{dplyr::mutate}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Apply dplyr's mutate to a subset of rows
}
