% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_design.R
\name{patchwork_design}
\alias{patchwork_design}
\title{Create a patchwork design specification}
\usage{
patchwork_design(..., margin.row = 0)
}
\arguments{
\item{...}{Character vectors specifying block dimensions. Each vector corresponds to a row in the output layout.

Each element of each vector is a character of the form \code{"WxH"}, (e.g. "8x10"). The \emph{height} of each row will be set to the height of the  tallest block in the row. Blocks with extra vertical space will be vertically centered. The \emph{width} of each row will be set to the width of the widest row, and any extra horizontal space will be evenly distributed in between blocks in the row.}

\item{margin.row}{An optional integer specifying the number of units of space to place in-between rows.}
}
\value{
A vector of patchwork areas.
}
\description{
Create a non-overlapping patchwork design layout by specifying a list of block dimensions.

\strong{Motivation:} When using the \code{"design"} argument of \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}, it can be cumbersome to use the syntax required by the patchwork framework. This syntax requires creating a vector of area elements, each of which is created by a call to \code{\link[patchwork:area]{patchwork::area()}}. This \code{area} function requires specifying the left/right/top/bottom coordinates of each block. This can become impractical when designing and making multiple edits to a complex layout in which we want our patches to be tiled/adjacent to one another. It can be much easier to instead specify a list of block dimensions, and have a function dynamically calculate all the necessary top/left/bottom/right coordinates. That is what this function does.
}
\examples{
# Create a layout with 2 blocks in the first row,
# 2 blocks in the second row, 1 block in the
# third row, and 2 blocks in the fourth row.

patchwork_design(
  c("8x10", "6.5x18"),
  c("9x12", "10x17"),
  c("4x28"),
  c("11x9", "11x13"),
  margin.row = 2
)
}
\seealso{
\code{\link[patchwork:area]{patchwork::area()}}, \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}
}
