% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sourceRmd.R
\name{read_chunk_rmd}
\alias{read_chunk_rmd}
\title{Read chunks from an Rmarkdown file}
\usage{
read_chunk_rmd(path, labels = NULL, prevent_overwrite = c("setup"))
}
\arguments{
\item{path}{Path to the R script.}

\item{labels}{Character vector of chunk labels (default \code{NULL}).}

\item{prevent_overwrite}{Character vector of chunk labels in the current document to prevent overwriting of; default is just the "setup" chunk.}
}
\description{
Extends the functionality of \code{\link[knitr:read_chunk]{knitr::read_chunk()}} to read from .Rmd files.
}
\seealso{
\code{\link[knitr:read_chunk]{knitr::read_chunk()}}, \code{\link[knitr:knit]{knitr::purl()}}
}
