% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_recurseRemoveNull.R
\name{remove_null}
\alias{remove_null}
\title{Remove NULL recursively}
\usage{
remove_null(x)
}
\arguments{
\item{x}{List-like object to recurse}
}
\description{
Recursively remove NULL objects from nested lists (e.g. ggplot themes)
}
