% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filesave.R
\name{resave}
\alias{resave}
\title{Resave}
\usage{
resave(
  ...,
  list = character(),
  file = stop("'file' must be specified"),
  compress = "xz",
  ascii = FALSE,
  version = NULL,
  envir = parent.frame(),
  eval.promises = TRUE,
  precheck = TRUE
)
}
\arguments{
\item{...}{the names of the objects to be saved (as symbols or
    character strings).}

\item{list}{a character vector (or \code{\link[base]{NULL}}) containing the names of objects to be
    saved.}

\item{file}{a (writable binary-mode) \link[base]{connection} or the name of the
    file where the data will be saved (when \link[base]{tilde expansion}
    is done).  Must be a file name for \code{save.image} or
    \code{version = 1}.}

\item{compress}{logical or character string specifying whether saving
    to a named file is to use compression.  \code{TRUE} corresponds to
    \command{gzip} compression, and character strings \code{"gzip"},
    \code{"bzip2"}, \code{"xz"} or \code{"zstd"} specify the type of
    compression.  Ignored when \code{file} is a connection and
    for workspace format version 1.}

\item{ascii}{if \code{TRUE}, an ASCII representation of the data is
    written.  The default value of \code{ascii} is \code{FALSE} which
    leads to a binary file being written.  If \code{NA} and
    \code{version >= 2}, a different ASCII representation is used which
    writes double/complex numbers as binary fractions.}

\item{version}{the workspace format version to use.  \code{NULL}
    specifies the current default format (3).  Version 1 was the default
    from \R 0.99.0 to \R 1.3.1 and version 2 from \R 1.4.0 to 3.5.0.
    Version 3 is supported from \R 3.5.0.}

\item{envir}{environment to search for objects to be saved.}

\item{eval.promises}{logical: should objects which are promises be
    forced before saving?}

\item{precheck}{logical: should the existence of the objects be
    checked before starting to save (and in particular before opening
    the file/connection)?  Does not apply to version 1 saves.}
}
\description{
Add objects to existing Rdata file
}
\seealso{
\code{\link[base:save]{base::save()}}
}
