% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_beautify.R
\name{sci_beautify}
\alias{sci_beautify}
\title{Beautify a number in scientific format}
\usage{
sci_beautify(
  n,
  sig_digits = 3,
  sci.format = "x",
  sci.mode = "auto",
  output.format = "plain",
  justify = F,
  padzeros = T
)
}
\arguments{
\item{n}{Number or numeric vector to convert}

\item{sig_digits}{(Optional) Number of significant digits (default = 3)}

\item{sci.format}{(Optional) Formatting of scientific notation. Can be one of "x" (multiplication character), "e" (lowercase e), "E" (uppercase E), "dot" (center dot). Default = "x".}

\item{sci.mode}{(Optional) Can be one of "auto" (automatically choose scientific notation or not), "on", or "off" (scientific notation always on or off).}

\item{output.format}{(Optional) Output target. Can be one of "plain", "html", or "latex". Default = "plain".}

\item{justify}{(Optional) Whether to left-justify string output across a character vector (default = FALSE)}

\item{padzeros}{(Optional) Whether to add trailing zeros (default = TRUE)}
}
\value{

}
\description{
Beautify a number in scientific format
}
