% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_beautify.R
\name{sci_notation_knit}
\alias{sci_notation_knit}
\title{Scientific notation format for knitr tables}
\usage{
sci_notation_knit(
  n,
  sig_digits = 3,
  output.format = NULL,
  sci.format = "x",
  sci.mode = "auto",
  justify = F,
  padzeros = T
)
}
\arguments{
\item{n}{Number or numeric vector to convert}

\item{sig_digits}{(Optional) Number of significant digits (default = 3)}

\item{output.format}{(Optional) Output target. Can be one of "plain", "html", or "latex". Default = "plain".}

\item{sci.format}{(Optional) Formatting of scientific notation. Can be one of "x" (multiplication character), "e" (lowercase e), "E" (uppercase E), "dot" (center dot). Default = "x".}

\item{sci.mode}{(Optional) Can be one of "auto" (automatically choose scientific notation or not), "on", or "off" (scientific notation always on or off).}

\item{justify}{(Optional) Whether to left-justify string output across a character vector (default = FALSE)}

\item{padzeros}{(Optional) Whether to add trailing zeros (default = TRUE)}
}
\value{

}
\description{
Takes as input a number (or vector of numbers), and outputs a character (or vector of characters) for use in a knitr table. This function will format the number according to the number of significant digits, and using the appropriate characters for scientific notation. By default, the function will check whether the knit output is html or graphic (LaTeX), and use the apporpriate string codes.
}
