% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_signifBeautify.R
\name{signif_beautify}
\alias{signif_beautify}
\title{Convert a number to a string in significant-figure form}
\usage{
signif_beautify(
  n,
  sig_digits,
  justify_mode,
  decimal_mark = ".",
  big_mark = ",",
  big_interval = 3,
  pad_zeros = F,
  justify_char = " "
)
}
\arguments{
\item{n}{Number or numeric vector to convert}

\item{sig_digits}{Number of significant digits. If NA or NULL, the input number is not changed.}

\item{justify_mode}{Mode to use for justifying a vector of values. Can be "l" (left), "c" (center), "r" (right), or "d" (decimal).}

\item{decimal_mark}{Character to use as the decimal marker.}

\item{big_mark}{If not blank, character to use in grouping every \code{big_interval} digits before the decimal marker.}

\item{big_interval}{Number of digits between each \code{big_mark}, if used.}

\item{pad_zeros}{Whether to add extra trailing zeros and/or a decimal to force number of significant digits (even if the original number did not have that many digits). Default = FALSE.}

\item{justify_char}{Character (or string) used to pad strings.}
}
\description{
Standard significant figure functions (such as \code{signif}, \code{format}) either do not output a string format (thereby excluding necessary trailing decimals & zeros), or do not correctly pad the number with the correct number of trailing zeros. This function aims to address these issues.
}
