% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sourceRmd.R
\name{sourceRmd}
\alias{sourceRmd}
\title{Source the R code from an RMarkdown file}
\usage{
sourceRmd(path, wd = dirname(path))
}
\arguments{
\item{path}{Path to RMarkdown file}

\item{wd}{Optional working directory to use for script execution. Defaults to the directory containing the RMarkdown file.}
}
\value{
A new environment containing the results of the sourced R code
}
\description{
Retrieve the R code from an RMarkdown file and source it into a new environment variable. Uses \code{\link[knitr:knit]{knitr::purl()}}, which comments-out any chunks that are not meant to be evaluated (i.e., \code{eval = F}).
}
