% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stringAddition.R
\name{str_plus}
\alias{str_plus}
\alias{\%paste\%}
\alias{\%paste\%.default}
\alias{\%paste\%.character}
\alias{\%slash\%}
\alias{\%slash\%.default}
\alias{\%slash\%.character}
\title{String concatenation}
\usage{
a \%paste\% b

\method{\%paste\%}{default}(a, b)

\method{\%paste\%}{character}(a, b)

a \%slash\% b

\method{\%slash\%}{default}(a, b)

\method{\%slash\%}{character}(a, b)
}
\arguments{
\item{a}{string a}

\item{b}{string b}
}
\description{
\%paste\% is for standard string concatenation. \%slash\% is for concatenating file paths.
}
