% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonStandardEval.R
\name{substitute_q}
\alias{substitute_q}
\title{A version of substitute that evaluates its first argument
This version of substitute is needed because substitute does not evaluate it's first argument, and it's often useful to be able to modify a quoted call.}
\usage{
substitute_q(call, env = parent.frame())
}
\arguments{
\item{call}{a quoted call}

\item{env}{an environment}
}
\description{
A version of substitute that evaluates its first argument
This version of substitute is needed because substitute does not evaluate it's first argument, and it's often useful to be able to modify a quoted call.
}
