% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrixOperations.R
\name{transpose}
\alias{transpose}
\title{Transpose a table preserving data.frame}
\usage{
transpose(data)
}
\arguments{
\item{data}{Tidy table}
}
\description{
Transpose a table preserving data.frame
}
