% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filesave.R
\name{write_parquet_zstd}
\alias{write_parquet_zstd}
\title{Write a zstd-compressed parquet file}
\usage{
write_parquet_zstd(
  x,
  sink,
  compression_level = 6,
  write_statistics = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame}, \link[arrow]{RecordBatch}, or \link[arrow]{Table}}

\item{sink}{A string file path, connection, URI, or \link[arrow]{OutputStream}, or path in a file
system (\code{SubTreeFileSystem})}

\item{compression_level}{compression level. Meaning depends on compression
algorithm}

\item{write_statistics}{logical: include statistics? Default \code{TRUE}}

\item{...}{
  Arguments passed on to \code{\link[arrow:write_parquet]{arrow::write_parquet}}
  \describe{
    \item{\code{chunk_size}}{how many rows of data to write to disk at once. This
directly corresponds to how many rows will be in each row group in
parquet. If \code{NULL}, a best guess will be made for optimal size (based on
the number of columns and number of rows), though if the data has fewer
than 250 million cells (rows x cols), then the total number of rows is
used.}
    \item{\code{version}}{parquet version: "1.0", "2.4" (default), "2.6", or
"latest" (currently equivalent to 2.6). Numeric values are
coerced to character.}
    \item{\code{compression}}{compression algorithm. Default "snappy". See details.}
    \item{\code{use_dictionary}}{logical: use dictionary encoding? Default \code{TRUE}}
    \item{\code{data_page_size}}{Set a target threshold for the approximate encoded
size of data pages within a column chunk (in bytes). Default 1 MiB.}
    \item{\code{use_deprecated_int96_timestamps}}{logical: write timestamps to INT96
Parquet format, which has been deprecated? Default \code{FALSE}.}
    \item{\code{coerce_timestamps}}{Cast timestamps a particular resolution. Can be
\code{NULL}, "ms" or "us". Default \code{NULL} (no casting)}
    \item{\code{allow_truncated_timestamps}}{logical: Allow loss of data when coercing
timestamps to a particular resolution. E.g. if microsecond or nanosecond
data is lost when coercing to "ms", do not raise an exception. Default
\code{FALSE}.}
  }}
}
\description{
A basic wrapper around \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.
}
