% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filesave.R
\name{write_parquet_zstd}
\alias{write_parquet_zstd}
\title{Write a zstd-compressed parquet file}
\usage{
write_parquet_zstd(
  x,
  sink,
  compression_level = 6,
  write_statistics = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame}, \link[arrow]{RecordBatch}, or \link[arrow]{Table}}

\item{sink}{A string file path, URI, or \link[arrow]{OutputStream}, or path in a file
system (\code{SubTreeFileSystem})}

\item{compression_level}{compression level. Meaning depends on compression algorithm}

\item{write_statistics}{Specify if we should write statistics. Default \code{TRUE}}

\item{...}{
  Arguments passed on to \code{\link[arrow:write_parquet]{arrow::write_parquet}}
  \describe{
    \item{\code{chunk_size}}{chunk size in number of rows. If NULL, the total number of rows is used.}
    \item{\code{version}}{parquet version, "1.0" or "2.0". Default "1.0". Numeric values
are coerced to character.}
    \item{\code{compression}}{compression algorithm. Default "snappy". See details.}
    \item{\code{use_dictionary}}{Specify if we should use dictionary encoding. Default \code{TRUE}}
    \item{\code{data_page_size}}{Set a target threshold for the approximate encoded
size of data pages within a column chunk (in bytes). Default 1 MiB.}
    \item{\code{use_deprecated_int96_timestamps}}{Write timestamps to INT96 Parquet format. Default \code{FALSE}.}
    \item{\code{coerce_timestamps}}{Cast timestamps a particular resolution. Can be
\code{NULL}, "ms" or "us". Default \code{NULL} (no casting)}
    \item{\code{allow_truncated_timestamps}}{Allow loss of data when coercing timestamps to a
particular resolution. E.g. if microsecond or nanosecond data is lost when coercing
to "ms", do not raise an exception}
    \item{\code{properties}}{A \code{ParquetWriterProperties} object, used instead of the options
enumerated in this function's signature. Providing \code{properties} as an argument
is deprecated; if you need to assemble \code{ParquetWriterProperties} outside
of \code{write_parquet()}, use \code{ParquetFileWriter} instead.}
    \item{\code{arrow_properties}}{A \code{ParquetArrowWriterProperties} object. Like
\code{properties}, this argument is deprecated.}
  }}
}
\description{
A basic wrapper around \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.
}
